restructure backend tests
