+++ backend/__tests__/helpers/testSetup.ts
@@ -1,0 +1,9 @@
+import db from '../../src/utils/db'
+
+beforeAll(async () => {
+  await db.connectToDatabase()
+})
+
+afterAll(async () => {
+  await db.sequelize.close()
+})
