--- backend/__tests__/recipeResolvers.test.ts
@@ -1,588 +1,0 @@
-import { existingCustomUserToken, existingUserToken, resetQuery, superuserToken, testServer } from './testHelper'
-import db from '../src/utils/db'
-import request from 'supertest'
-import { Server } from 'http'
-
-let httpServer: Server
-let token: string
-
-const createRecipeQuery = /* GraphQL */`
-  mutation createRecipe(
-    $name: String!
-    $description: String
-    $ingredientCategories: [IngredientCategoryInput!]!
-    $steps: [String!]!,
-    $serving: ServingInput,
-    $prepareTime: Int
-  ) {
-  createRecipe (
-    name: $name
-    description: $description
-    ingredientCategories: $ingredientCategories
-    steps: $steps
-    serving: $serving
-    prepareTime: $prepareTime
-  ) {
-    id
-    name
-    description
-    totalLikes
-    likedByCurrentUser
-    ingredientCategories {
-      name
-      ingredients {
-        amount
-        name
-        unit
-      }
-    }
-    steps
-    serving {
-      amount
-      per
-      unit
-    }
-    prepareTime
-    user {
-      id
-      username
-    }
-  }
-}
-`
-
-const recipeLikesQuery = /* GraphQL */`
-  query recipe( $id: ID! ) { recipe ( id: $id ) { totalLikes likedByCurrentUser } }
-`
-
-const deleteRecipQuery = /* GraphQL */`
-  mutation deleteRecipe( $id: ID! ) { deleteRecipe ( id: $id ) }
-`
-
-const recipesQuery = /* GraphQL */`
-  query recipes( $search: String ) { recipes ( page: 0, search: $search ) { count rows { name } } }
-`
-
-const likeRecipeQuery = /* GraphQL */`
-  mutation likeRecipe( $id: ID! ) { likeRecipe ( id: $id ) { recipeId, userId }}
-`
-
-const removeRecipeLikeQuery = /* GraphQL */`
-  mutation removeRecipeLike( $id: ID! ) { removeRecipeLike ( id: $id ) }
-`
-
-const testVariables = {
-  name: 'breakfast porridge in the microwave',
-  ingredientCategories: [{
-    name: 'porridge',
-    ingredients: [{
-      name: 'four grain flakes',
-      amount: 1,
-      unit: 'dl',
-    }, {
-      name: 'water',
-      amount: 2,
-      unit: 'dl',
-    }],
-  }, {
-    name: 'toppings',
-    ingredients: [{
-      name: 'salt',
-      amount: 0.5,
-      unit: 'tsp',
-    }, {
-      name: 'butter',
-      amount: 25,
-      unit: 'g',
-    },
-    {
-      name: 'fresh berries',
-      amount: 50,
-      unit: 'g',
-    }],
-  }],
-  steps: [
-    'Mix porridge ingredients in a bowl',
-    'Microwave at 800W for 3 minutes',
-    'Top with salt and butter or berries',
-  ],
-}
-
-const testVariablesFull = {
-  ...testVariables,
-  description: 'high fiber porridge for breakfast',
-  serving: {
-    amount: 1,
-    per: 3,
-    unit: 'dl',
-  },
-  prepareTime: 5,
-}
-
-describe('recipeResolver tests', () => {
-  beforeAll(async () => {
-    await db.connectToDatabase()
-    httpServer = await testServer
-    await request(httpServer).post('/').send({ query: resetQuery })
-  })
-
-  afterAll(async () => {
-    await db.sequelize.close()
-  })
-
-  describe('creating recipe', () => {
-    beforeEach(async () => {
-      await request(httpServer).post('/').send({ query: resetQuery })
-      token = await existingUserToken(httpServer)
-    })
-
-    test('succeeds with correct input', async () => {
-      const variables = testVariablesFull
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      expect(response.body.errors).toBeUndefined()
-      expect(response.body.data.createRecipe.name).toBe('breakfast porridge in the microwave')
-    })
-
-    test('returns correct fields after success', async () => {
-      const variables = testVariablesFull
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      expect(response.body.errors).toBeUndefined()
-      const data = response.body.data.createRecipe
-
-      expect(data.id).not.toBeNull()
-      expect(data.name).toEqual(testVariablesFull.name)
-      expect(data.description).toEqual(testVariablesFull.description)
-      expect(data.ingredientCategories).toEqual(testVariablesFull.ingredientCategories)
-      expect(data.prepareTime).toEqual(testVariablesFull.prepareTime)
-      expect(data.serving).toEqual(testVariablesFull.serving)
-      expect(data.steps).toEqual(testVariablesFull.steps)
-      expect(data.user.id).not.toBeNull()
-      expect(data.user.username).not.toBeNull()
-      expect(data.totalLikes).toEqual(0)
-      expect(data.likedByCurrentUser).toEqual(false)
-    })
-
-    test('succeeds with only necessary fields', async () => {
-      const variables = testVariables
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      expect(response.body.errors).toBeUndefined()
-      expect(response.body.data.createRecipe.name).toBe('breakfast porridge in the microwave')
-    })
-
-    test('fails without auth token', async () => {
-      const variables = testVariables
-      const response = await request(httpServer).post('/').send({ query: createRecipeQuery, variables })
-
-      expect(response.body.errors[0].message).toBe('Unauthorized')
-    })
-
-    test('fails without recipe name', async () => {
-      const variables = { ...testVariables, name: null }
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      expect(response.body.errors[0].extensions.code).toBe('BAD_USER_INPUT')
-    })
-
-    test('fails without recipe steps', async () => {
-      const variables = { ...testVariables, steps: [] }
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      expect(response.body.errors[0].extensions.code).toBe('BAD_USER_INPUT')
-    })
-
-    test('fails without recipe ingredients', async () => {
-      const variables = { ...testVariables, ingredientCategories: [] }
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      expect(response.body.errors[0].extensions.code).toBe('BAD_USER_INPUT')
-    })
-
-    test('fails with too many steps', async () => {
-      const variables = { ...testVariables, steps: ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11'] }
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      expect(response.body.errors[0].extensions.code).toBe('BAD_USER_INPUT')
-    })
-  })
-
-  describe('deleting recipe', () => {
-    beforeEach(async () => {
-      await request(httpServer).post('/').send({ query: resetQuery })
-      token = await existingUserToken(httpServer)
-    })
-
-    test('succeeds with same user', async () => {
-      const variables = testVariables
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: deleteRecipQuery, variables: { id: recipeId } })
-        .set({ Authorization: token })
-
-      expect(response2.body.errors).toBeUndefined()
-      expect(response2.body.data.deleteRecipe).toBe(true)
-    })
-
-    test('fails without auth token', async () => {
-      const variables = testVariables
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: deleteRecipQuery, variables: { id: recipeId } })
-
-      expect(response2.body.errors[0].message).toBe('Unauthorized')
-      expect(response2.body.data.deleteRecipe).not.toBe(true)
-    })
-
-    test('fails with wrong token', async () => {
-      const variables = testVariables
-      const token2 = await existingCustomUserToken(httpServer, 'TestUser2', 'Password1')
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: deleteRecipQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      expect(response2.body.errors[0].extensions.code).toBe('BAD_USER_INPUT')
-      expect(response2.body.data.deleteRecipe).not.toBe(true)
-    })
-
-    test('succeeds with superuser', async () => {
-      const variables = testVariables
-      const superToken = await superuserToken(httpServer)
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: deleteRecipQuery, variables: { id: recipeId } })
-        .set({ Authorization: superToken })
-
-      expect(response2.body.errors).toBeUndefined()
-      expect(response2.body.data.deleteRecipe).toBe(true)
-    })
-  })
-
-  describe('searching recipes', () => {
-    beforeEach(async () => {
-      await request(httpServer).post('/').send({ query: resetQuery })
-      token = await existingUserToken(httpServer)
-    })
-
-    test('succeeds with exact search', async () => {
-      const variables = testVariables
-      await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: recipesQuery, variables: { search: 'breakfast porridge in the microwave' } })
-
-      expect(response.body.data.recipes.rows).toHaveLength(1)
-    })
-
-    test('succeeds with partial search', async () => {
-      const variables = testVariables
-      await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: recipesQuery, variables: { search: 'PORRIDGE' } })
-
-      expect(response.body.data.recipes.rows).toHaveLength(1)
-    })
-
-    test('fails with invalid search', async () => {
-      const variables = testVariables
-      await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const response = await request(httpServer)
-        .post('/')
-        .send({ query: recipesQuery, variables: { search: 'macaroni and cheese' } })
-
-      expect(response.body.data.recipes.rows).toHaveLength(0)
-    })
-  })
-  describe('liking recipe', () => {
-    beforeEach(async () => {
-      await request(httpServer).post('/').send({ query: resetQuery })
-      token = await existingUserToken(httpServer)
-    })
-    test('succeeds with existing recipe', async () => {
-      const variables = testVariables
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token })
-
-      expect(response2.body.errors).toBeUndefined()
-      expect(response2.body.data.likeRecipe.recipeId).toBe(recipeId)
-    })
-    test('fails without auth token', async () => {
-      const variables = testVariables
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-
-      expect(response2.body.errors[0].extensions.code).toBe('BAD_USER_INPUT')
-    })
-    test('fails with invalid id', async () => {
-      const variables = testVariables
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id as number
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId + 1 } })
-        .set({ Authorization: token })
-
-      expect(response2.body.errors[0].extensions.code).toBe('BAD_USER_INPUT')
-    })
-    test('succeeds with different user', async () => {
-      const variables = testVariables
-      const token2 = await existingCustomUserToken(httpServer, 'TestUser2', 'Password1')
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      expect(response2.body.data.likeRecipe.recipeId).toBe(recipeId)
-    })
-    test('shows correctly with multiple likes', async () => {
-      const variables = testVariables
-      const token2 = await existingCustomUserToken(httpServer, 'TestUser2', 'Password1')
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: recipeLikesQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      expect(response2.body.data.recipe.totalLikes).toEqual(2)
-      expect(response2.body.data.recipe.likedByCurrentUser).toEqual(true)
-    })
-    test('is not liked by currentUser if liked by others', async () => {
-      const variables = testVariables
-      const token2 = await existingCustomUserToken(httpServer, 'TestUser2', 'Password1')
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token })
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: recipeLikesQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      expect(response2.body.data.recipe.totalLikes).toEqual(1)
-      expect(response2.body.data.recipe.likedByCurrentUser).toEqual(false)
-    })
-    test('is not liked by currentUser if not logged in', async () => {
-      const variables = testVariables
-      const token2 = await existingCustomUserToken(httpServer, 'TestUser2', 'Password1')
-
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token2 })
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: recipeLikesQuery, variables: { id: recipeId } })
-
-      expect(response2.body.data.recipe.totalLikes).toEqual(2)
-      expect(response2.body.data.recipe.likedByCurrentUser).toEqual(false)
-    })
-  })
-  describe('removing recipe like', () => {
-    beforeEach(async () => {
-      await request(httpServer).post('/').send({ query: resetQuery })
-      token = await existingUserToken(httpServer)
-    })
-    test('succeeds when like exists', async () => {
-      const variables = testVariables
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token })
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: removeRecipeLikeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token })
-
-      expect(response2.body.errors).toBeUndefined()
-      expect(response2.body.data.removeRecipeLike).toEqual(true)
-    })
-    test('fails without auth token', async () => {
-      const variables = testVariables
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token })
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: removeRecipeLikeQuery, variables: { id: recipeId } })
-
-      expect(response2.body.errors[0].message).toBe('Unauthorized')
-      expect(response2.body.data.removeRecipeLike).not.toEqual(true)
-    })
-    test('fails with invalid id', async () => {
-      const variables = testVariables
-      const response1 = await request(httpServer)
-        .post('/')
-        .send({ query: createRecipeQuery, variables })
-        .set({ Authorization: token })
-
-      const recipeId = response1.body.data.createRecipe.id as number
-
-      await request(httpServer)
-        .post('/')
-        .send({ query: likeRecipeQuery, variables: { id: recipeId } })
-        .set({ Authorization: token })
-
-      const response2 = await request(httpServer)
-        .post('/')
-        .send({ query: removeRecipeLikeQuery, variables: { id: recipeId + 1 } })
-        .set({ Authorization: token })
-
-      expect(response2.body.errors[0].message).toBe('Like not found')
-      expect(response2.body.data.removeRecipeLike).not.toEqual(true)
-    })
-  })
-})
